split lib.util to several sub-modules, for clarity
