+++ lib/util/__init__.py
@@ -1,0 +1,16 @@
+import inspect
+from pathlib import Path
+
+# directory with all these modules, and potentially more files
+# - useful until TMT can parametrize 'environment:' with variable expressions,
+#   so we could add the libdir to PATH and PYTHONPATH
+# TODO: after RHEL-7, replace with importlib.resources to access files
+#       in the python package hierarchy, python 3.7+
+libdir = Path(inspect.getfile(inspect.currentframe())).parent.parent
+
+from .content      import *  # noqa
+from .environment  import *  # noqa
+from .httpsrv      import *  # noqa
+from .log          import *  # noqa
+from .sanitization import *  # noqa
+from .subprocess   import *  # noqa
