+++ lib/util/environment.py
@@ -1,0 +1,16 @@
+import os
+
+
+def running_in_beaker():
+    """
+    Return True if running in Beaker under the FMF wrapper.
+    """
+    taskpath = os.environ.get('RSTRNT_TASKPATH')
+    return bool(taskpath and taskpath.endswith('/distribution/wrapper/fmf'))
+
+
+def running_in_tmt():
+    """
+    Return True if running under TMT.
+    """
+    return bool(os.environ.get('TMT_TEST_DATA'))
